move version and path related functions to new lib
